pyproject.toml
[project]
name = "git-commit-gen"
version = "0.2.0"
description = "Generate Conventional Commit messages from staged git diff using OpenAI."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.3",
    "PyYAML>=6.0.2",
]

[dependency-groups]
dev = [
    "pyinstaller>=6.12.0",
]

[tool.uv]
package = false
